fix protect auth

+++ app/Http/Controllers/EventOneController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\FormOneRequest;
 use App\Models\EventOne;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class EventOneController extends Controller
@@ -19,49 +20,62 @@ class EventOneController extends Controller
      */
     public function index()
     {
-        $total_enrollments = EventOne::all()
-            ->count();
-        $total_confirmed_enrollments = EventOne::all()
-            ->where('email_verified_at','!=',null )
-            ->count();
-        $total_not_confirmed_enrollments = EventOne::all()
-            ->where('email_verified_at','=',null )
-            ->count();
-        $inscritos = EventOne::all()
-            ->where('email_verified_at','!=', null);
+        if (Auth::check()) {
+            // The user is logged in...
+
+            $total_enrollments = EventOne::all()
+                ->count();
+            $total_confirmed_enrollments = EventOne::all()
+                ->where('email_verified_at','!=',null )
+                ->count();
+            $total_not_confirmed_enrollments = EventOne::all()
+                ->where('email_verified_at','=',null )
+                ->count();
+            $inscritos = EventOne::all()
+                ->where('email_verified_at','!=', null);
 
 
-        return view('applications.eventone.index', [
-            'total_enrollments' => $total_enrollments,
-            'total_confirmed_enrollments'=>$total_confirmed_enrollments,
-            'total_not_confirmed_enrollments'=>$total_not_confirmed_enrollments,
-            'inscritos'=>$inscritos
-        ]);
+            return view('applications.eventone.index', [
+                'total_enrollments' => $total_enrollments,
+                'total_confirmed_enrollments'=>$total_confirmed_enrollments,
+                'total_not_confirmed_enrollments'=>$total_not_confirmed_enrollments,
+                'inscritos'=>$inscritos
+            ]);
+        }
+        return view('auth.login');
+
     }
 
     public function not_confirmed()
     {
-        $total_enrollments = EventOne::all()
-            ->count();
-        $total_confirmed_enrollments = EventOne::all()
-            ->where('email_verified_at','!=',null )
-            ->count();
-        $total_not_confirmed_enrollments = EventOne::all()
-            ->where('email_verified_at','=',null )
-            ->count();
-        $inscritos = EventOne::all()
-            ->where('email_verified_at','!=', null);
-        $not_confirmed = EventOne::all()
-            ->where('email_verified_at','=', null);
+        if (Auth::check()) {
+            // The user is logged in...
+
+            $total_enrollments = EventOne::all()
+                ->count();
+            $total_confirmed_enrollments = EventOne::all()
+                ->where('email_verified_at','!=',null )
+                ->count();
+            $total_not_confirmed_enrollments = EventOne::all()
+                ->where('email_verified_at','=',null )
+                ->count();
+            $inscritos = EventOne::all()
+                ->where('email_verified_at','!=', null);
+            $not_confirmed = EventOne::all()
+                ->where('email_verified_at','=', null);
 
 
-        return view('applications.eventone.not_confirmed', [
-            'total_enrollments' => $total_enrollments,
-            'total_confirmed_enrollments'=>$total_confirmed_enrollments,
-            'total_not_confirmed_enrollments'=>$total_not_confirmed_enrollments,
-            'inscritos'=>$inscritos,
-            'not_confirmed'=>$not_confirmed
-        ]);
+            return view('applications.eventone.not_confirmed', [
+                'total_enrollments' => $total_enrollments,
+                'total_confirmed_enrollments'=>$total_confirmed_enrollments,
+                'total_not_confirmed_enrollments'=>$total_not_confirmed_enrollments,
+                'inscritos'=>$inscritos,
+                'not_confirmed'=>$not_confirmed
+            ]);
+        }
+        return view('auth.login');
+
+
     }
 
 
@@ -112,18 +126,22 @@ class EventOneController extends Controller
 
     public function manual_confirmed($uuid)
     {
-        $eventOne = EventOne::findByUuid($uuid);
-        $eventOne->update(['email_verified_at' => now()]);
-        $eventOne->save();
+        if (Auth::check()) {
+            // The user is logged in...
+            $eventOne = EventOne::findByUuid($uuid);
+            $eventOne->update(['email_verified_at' => now()]);
+            $eventOne->save();
 
-        toastr()->success("Inscrição confirmada manualmente com sucesso!.",
-            'Atenção - O registro de verificação foi definido manualmente',
-            ['closeButton'=>true,
-                'positionClass'=>'toast-top-right',
-                'timeOut'=>'4000'
-            ]);
+            toastr()->success("Inscrição confirmada manualmente com sucesso!.",
+                'Atenção - O registro de verificação foi definido manualmente',
+                ['closeButton'=>true,
+                    'positionClass'=>'toast-top-right',
+                    'timeOut'=>'4000'
+                ]);
 
-        return redirect()->back();
+            return redirect()->back();
+        }
+
 
     }
 
@@ -146,23 +164,28 @@ class EventOneController extends Controller
      */
     public function edit(EventOne $eventOne)
     {
-        $total_enrollments = EventOne::all()
-            ->count();
-        $total_confirmed_enrollments = EventOne::all()
-            ->where('email_verified_at','!=',null )
-            ->count();
-        $total_not_confirmed_enrollments = EventOne::all()
-            ->where('email_verified_at','=',null )
-            ->count();
-        $inscritos = EventOne::all()
-            ->where('email_verified_at','!=', null);
-        return view('applications.eventone.edit',[
-            'eventOne'=>$eventOne,
-            'total_enrollments' => $total_enrollments,
-            'total_confirmed_enrollments'=>$total_confirmed_enrollments,
-            'total_not_confirmed_enrollments'=>$total_not_confirmed_enrollments,
-            'inscritos'=>$inscritos
-        ]);
+        if (Auth::check()) {
+            // The user is logged in...
+            $total_enrollments = EventOne::all()
+                ->count();
+            $total_confirmed_enrollments = EventOne::all()
+                ->where('email_verified_at','!=',null )
+                ->count();
+            $total_not_confirmed_enrollments = EventOne::all()
+                ->where('email_verified_at','=',null )
+                ->count();
+            $inscritos = EventOne::all()
+                ->where('email_verified_at','!=', null);
+            return view('applications.eventone.edit',[
+                'eventOne'=>$eventOne,
+                'total_enrollments' => $total_enrollments,
+                'total_confirmed_enrollments'=>$total_confirmed_enrollments,
+                'total_not_confirmed_enrollments'=>$total_not_confirmed_enrollments,
+                'inscritos'=>$inscritos
+            ]);
+
+        }
+
     }
 
     /**
@@ -174,31 +197,35 @@ class EventOneController extends Controller
      */
     public function update(FormOneRequest $request, EventOne $eventOne)
     {
-        $eventOne->update($request->validated());
+        if (Auth::check()) {
+            // The user is logged in...
+            $eventOne->update($request->validated());
 
-        if ($eventOne->wasChanged())
-        {
+            if ($eventOne->wasChanged())
+            {
 
 
-            toastr()->success("Inscrição Atualizada com sucesso!.",
-                'Atenção - As Informações sobre a Inscrição foram Modificadas!!!!',
-                ['closeButton'=>true,
-                    'positionClass'=>'toast-top-right',
-                    'timeOut'=>'4000'
-                ]);
+                toastr()->success("Inscrição Atualizada com sucesso!.",
+                    'Atenção - As Informações sobre a Inscrição foram Modificadas!!!!',
+                    ['closeButton'=>true,
+                        'positionClass'=>'toast-top-right',
+                        'timeOut'=>'4000'
+                    ]);
+            }
+
+
+            else
+                toastr()->info("Inscrição Não foi Alterada!.",
+                    'Atenção - Não houve Modificações!!!!',
+                    ['closeButton'=>true,
+                        'positionClass'=>'toast-top-right',
+                        'timeOut'=>'4000'
+                    ]);
+
+
+            return redirect()->route('event_one.index');
         }
 
-
-        else
-            toastr()->info("Inscrição Não foi Alterada!.",
-                'Atenção - Não houve Modificações!!!!',
-                ['closeButton'=>true,
-                    'positionClass'=>'toast-top-right',
-                    'timeOut'=>'4000'
-                ]);
-
-
-        return redirect()->route('event_one.index');
 
 
     }
@@ -213,14 +240,20 @@ class EventOneController extends Controller
      */
     public function destroy(EventOne $eventOne)
     {
-        $eventOne->delete();
-        toastr()->error("Inscrição DESTRUÍDA!.",
-            'Atenção - Registro foi destruído!!!!',
-            ['closeButton'=>true,
-                'positionClass'=>'toast-top-right',
-                'timeOut'=>'4000'
-            ]);
-        return redirect()->back();
+        if (Auth::check()) {
+            // The user is logged in...
+
+            $eventOne->delete();
+            toastr()->error("Inscrição DESTRUÍDA!.",
+                'Atenção - Registro foi destruído!!!!',
+                ['closeButton'=>true,
+                    'positionClass'=>'toast-top-right',
+                    'timeOut'=>'4000'
+                ]);
+            return redirect()->back();
+        }
+
+
     }
